Refactor turn-countdown into separate component
